🏗️ Architect: Modularized planetary searches
Split the 570-line `apts/skyfield_searches/planets.py` into a modular package `apts/skyfield_searches/planets/`.

- `extrema.py`: Observational limits (dichotomy, stationary points, greatest elongations, highest altitude).
- `orbits.py`: Orbital events (aphelion, perihelion).
- `alignments.py`: Multi-body events (alignments, oppositions, Mars closest approach).

Maintained the public API via `__init__.py` re-exports and updated relative imports to use `...` for root-level modules. verified with existing unit tests.
